More functional keepalive configuration
When keepalive is enabled, Manticore (Apache Commons HTTP) may
reuse connections that were closed by the remote server. It also
doesn't retry non-idempotent connections (e.g. POST) by default.

Unfortunately, we use POST to send logs (because otherwise, we wouldn't
be able to tell the difference between an incoming log record and a
healtcheck!).

In practice, this means that if there's a gap between two log records
that's longer than the keepalive, the next log record will be lost,
because it'll be sent over a dead connection and not retried.

To avoid this problem, we:

+ Enable `stale_check`, to avoid reusing dead connections in the first
  place
+ Enable `retry_non_idempotent`, so that if a dead connections is in fact
  reused, the message will be re-sent (the Commons doc explain that
  `stale_check` is not entirely reliable).